+++ components/ehmtxv2/__init__.py
@@ -77,6 +77,7 @@ CONF_RAINBOWINTERVAL = "rainbow_interval"
 CONF_FRAMEINTERVAL = "frame_interval"
 CONF_DEFAULT_FONT_ID = "default_font_id"
 CONF_DEFAULT_FONT = "default_font"
+CONF_CLOCKFONT = "default_clock_font"
 CONF_DEFAULT_FONT_XOFFSET = "default_font_xoffset"
 CONF_DEFAULT_FONT_YOFFSET = "default_font_yoffset"
 CONF_SPECIAL_FONT_ID = "special_font_id"
@@ -106,6 +107,9 @@ EHMTX_SCHEMA = cv.Schema({
     ): cv.templatable(cv.positive_int),
     cv.Optional(
         CONF_HTML, default=False
+    ): cv.boolean,
+    cv.Optional(
+        CONF_CLOCKFONT, default=True
     ): cv.boolean,
      cv.Optional(
         CONF_RTL, default=False
@@ -365,6 +369,7 @@ async def to_code(config):
 
     cg.add(var.set_brightness(config[CONF_BRIGHTNESS]))
     cg.add(var.set_scroll_interval(config[CONF_SCROLLINTERVAL]))
+    cg.add_define("EHMTXv2_SCROLL_INTERVALL",config[CONF_SCROLLINTERVAL])    
     cg.add(var.set_rainbow_interval(config[CONF_SCROLLINTERVAL]))
     if config[CONF_RTL]:
         cg.add(var.set_rtl(config[CONF_RTL]))
@@ -373,6 +378,10 @@ async def to_code(config):
     cg.add(var.set_frame_interval(config[CONF_FRAMEINTERVAL]))
     cg.add(var.set_week_start(config[CONF_WEEK_START_MONDAY]))
     cg.add(var.set_clock_interval(config[CONF_CLOCKINTERVAL]))
+    if config[CONF_CLOCKFONT]:
+        cg.add_define("EHMTXv2_DEFAULT_CLOCK_FONT","true")    
+    else:
+        cg.add_define("EHMTXv2_DEFAULT_CLOCK_FONT","false")    
     cg.add(var.set_time_format(config[CONF_TIME_FORMAT]))
     cg.add(var.set_date_format(config[CONF_DATE_FORMAT]))
     cg.add(var.set_show_day_of_week(config[CONF_SHOWDOW]))